update entry point for whats new modal
